requirements.txt
beautifulsoup4==4.10.0
ffmpeg-python==0.2.0
lxml==4.7.1
pandas==1.3.1
pdfplumber==0.6.0
python-docx==0.8.11
python-pptx==0.6.21
requests==2.25.1
tqdm==4.62.0
Unidecode==1.3.3
-e .